+++ src/ReaLTaiizor/Controls/TrackBar/PoisonTrackBar.cs
@@ -486,12 +486,12 @@ namespace ReaLTaiizor.Controls
             {
                 case Keys.Down:
                 case Keys.Left:
-                    SetProperValue(Value - (int)smallChange);
+                    SetProperValue(Value - smallChange);
                     OnScroll(ScrollEventType.SmallDecrement, Value);
                     break;
                 case Keys.Up:
                 case Keys.Right:
-                    SetProperValue(Value + (int)smallChange);
+                    SetProperValue(Value + smallChange);
                     OnScroll(ScrollEventType.SmallIncrement, Value);
                     break;
                 case Keys.Home:
@@ -501,11 +501,11 @@ namespace ReaLTaiizor.Controls
                     Value = barMaximum;
                     break;
                 case Keys.PageDown:
-                    SetProperValue(Value - (int)largeChange);
+                    SetProperValue(Value - largeChange);
                     OnScroll(ScrollEventType.LargeDecrement, Value);
                     break;
                 case Keys.PageUp:
-                    SetProperValue(Value + (int)largeChange);
+                    SetProperValue(Value + largeChange);
                     OnScroll(ScrollEventType.LargeIncrement, Value);
                     break;
             }
@@ -578,7 +578,7 @@ namespace ReaLTaiizor.Controls
                 Point pt = e.Location;
                 int p = pt.X;
 
-                float coef = (float)(barMaximum - barMinimum) / (float)(ClientSize.Width - 3);
+                float coef = (barMaximum - barMinimum) / (float)(ClientSize.Width - 3);
                 trackerValue = (int)(p * coef + barMinimum);
 
                 if (trackerValue <= barMinimum)
